Add files via upload
This is a little website of lyrics for a song called "Fairytale". "Fairytale" is a song by the Belarusian-Norwegian singer and violinist Alexander Rybak with which he won the Eurovision Song Contest in 2009 representing Norway. This site is a tribute to the phenomenal singer and violinist, Alexander Rybak. Created for Assignment of Interface Design using CSS, Georgian College, ON, Canada.